% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSFA_gibbs_run_chain.R
\name{NSFA_gibbs}
\alias{NSFA_gibbs}
\title{Gibbs Sampler for the Non-parametric Sparse Factor Analysis (NSFA) Model}
\usage{
NSFA_gibbs(S, Y, hp = NULL, Y_test = NULL, R = 1L, verbose = 0L)
}
\arguments{
\item{S}{number of iterations for the chain.}

\item{Y}{matrix with the observed data, of size \code{D x N}.}

\item{hp}{list of hyperparameters. See Details for a description and default
values used.}

\item{Y_test}{matrix with test set for assessing out-of-sample performance.
Number of columns must be same as \code{Y} (\code{N}).}

\item{R}{function returns last \code{R} values in the chain. Default is 1.}

\item{verbose}{integer. Function prints every \code{verbose} iterations. The
default value of 0 prints no output.}
}
\value{
List with two lists:
\describe{
  \item{chain}{list with last \code{R} samples of the chain.}
  \item{stats}{list of traces of train set log-likelihood (ll), test set ll, and K.}
}
}
\description{
This function runs a Gibbs sampler to obtain posterior samples from the NSFA
model by Knowles & Ghahramani (2011). This is my personal implementation, which
contains some corrections with respect to the model presented in the original
article.
}
\details{
The model was originally proposed in Knowles & Ghahramani (2011). However, you
can read my own report evaluating this model
\href{https://miguelbiron.github.io/docs/STAT548_report_3_NSFA.pdf}{here}.

The following hyperparameters can be supplied by the user through \code{hp}.
Default values are in parenthesis and tend to work well.
\itemize{
  \item \code{e}: shape parameter of prior for alpha (1).
  \item \code{f}: rate parameter of prior for alpha (1).
  \item \code{a}: shape parameter of prior for psi^{-1} (1).
  \item \code{a_0}: shape parameter of prior for b (1).
  \item \code{b_0}: rate parameter of prior for b (1).
  \item \code{c}: shape parameter of prior for lambda (1).
  \item \code{c_0}: shape parameter of prior for d (1).
  \item \code{d_0}: rate parameter of prior for d (1).
  \item \code{lambda_MH}: lambda parameter for kappa_d proposal (10).
  \item \code{pi_MH}: pi parameter for kappa_d proposal (0.1).
}

If chain keeps restarting, try increasing \code{lambda_MH} and/or \code{pi_MH}
to get more aggressive proposals for growing Z.
}
\references{
Knowles, D., & Ghahramani, Z. (2011). Nonparametric Bayesian sparse factor
models with application to gene expression modeling.
\emph{The Annals of Applied Statistics}, 1534-1552.
}
