% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBLM_run_chain.R
\name{sblm_gibbs}
\alias{sblm_gibbs}
\title{Gibbs sampler for a sparse Bayesian linear regression model}
\usage{
sblm_gibbs(X, y, hp = NULL, S, verbose = 0L)
}
\arguments{
\item{X}{design matrix, of size \code{N x P}.}

\item{y}{response vector of length \code{N}.}

\item{hp}{list of hyperparameters. See Details for a description and default
values used.}

\item{S}{number of iterations for the chain.}

\item{verbose}{integer. Function prints every \code{verbose} iterations. The
default value of 0 prints no output.}
}
\value{
A nested list with \code{S} elements, each one being a list of the
values of the latent variables at every iteration.
}
\description{
This function runs a Gibbs sampler for a Bayesian linear regression model
that explicitly allows for sparse solutions, in the spirit of the spike and
slab prior (Mitchell and Beauchamp 1988). The difference with this model
is that, instead of putting a prior on the coefficients with a point mass at
0, we use binary auxiliary variables to mask the effect of a variable from
the predicted values. More details about the model can be found in
\href{https://miguelbiron.github.io/}{UPDATE THIS LINK WHEN POST IS UP}.
}
\details{
The following hyperparameters can be supplied by the user through \code{hp}.
Default values for all are 1, and tend to work well.
\itemize{
  \item \code{a_0_b}: shape parameter of Inv-Gamma prior for sigma_2_b.
  \item \code{b_0_b}: scale parameter of Inv-Gamma prior for sigma_2_b.
  \item \code{a_0_e}: shape parameter of Inv-Gamma prior for sigma_2_e.
  \item \code{b_0_e}: scale parameter of Inv-Gamma prior for sigma_2_e.
  \item \code{s_1_pi}: shape 1 parameter for Beta prior of pi_z.
  \item \code{s_2_pi}: shape 2 parameter for Beta prior of pi_z.
}

Note that the default values for the prior of pi_z imply a uniform distribution
on (0,1). If the number of columns in \code{X} is large, and you suspect that
beta should be highly sparse, then reflecting this in the prior of pi_z would
speed up computation considerably.
}
\references{
Mitchell, Toby J, and John J Beauchamp. 1988. Bayesian Variable Selection in
Linear Regression. \emph{Journal of the American Statistical Association,
83} (404). Taylor & Francis Group: 1023–32.
}
